# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Kartala Task Management API",
        "title": "Kartala Task Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/api/v1/users/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get profile information for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user info",
                "responses": {
                    "200": {
                        "description": "User info retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUserMeResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerNotFoundResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 404
                        },
                        "message": {
                            "type": "string",
                            "example": "Resource not found"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
...
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUnauthorizedResponse": {
            "type": "object",
            "properties": {
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 401
                        },
                        "message": {
                            "type": "string",
                            "example": "Invalid or expired session"
                        },
                        "status": {
                            "type": "string",
                            "example": "error"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUserMeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUserMeResponseData"
                },
                "meta": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "integer",
                            "example": 200
                        },
                        "message": {
                            "type": "string",
                            "example": "Data fetched successfully"
                        },
                        "status": {
                            "type": "string",
                            "example": "success"
                        }
                    }
                }
            }
        },
        "github_com_adriannbhp_kartala-golang-task-management-api_pkg_response.SwaggerUserMeResponseData": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and then your token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}